fix(vulndb): split file locks by platform
