+++ internal/vulndb/filestore.go
@@ -9,7 +9,6 @@ import (
 	"path/filepath"
 	"sort"
 	"sync"
-	"syscall"
 	"time"
 )
 
@@ -272,25 +271,6 @@ func (s *FileStore) read(ctx context.Context, apply func() error) error {
 		return err
 	}
 	return apply()
-}
-
-func (s *FileStore) lockStateFile() (func(), error) {
-	dir := filepath.Dir(s.path)
-	if err := os.MkdirAll(dir, 0o700); err != nil {
-		return nil, err
-	}
-	lock, err := os.OpenFile(s.path+".lock", os.O_CREATE|os.O_RDWR, 0o600)
-	if err != nil {
-		return nil, err
-	}
-	if err := syscall.Flock(int(lock.Fd()), syscall.LOCK_EX); err != nil {
-		_ = lock.Close()
-		return nil, err
-	}
-	return func() {
-		_ = syscall.Flock(int(lock.Fd()), syscall.LOCK_UN)
-		_ = lock.Close()
-	}, nil
 }
 
 func (s *FileStore) saveLocked(ctx context.Context) error {
